# fix randomization